+++ src/gametest/java/dev/kkayam/infinitestack/gametest/TestCommand.java
@@ -1,4 +1,4 @@
-package dev.kkayam.storagesolution.gametest;
+package dev.kkayam.infinitestack.gametest;
 
 import com.mojang.brigadier.arguments.StringArgumentType;
 import net.fabricmc.api.ModInitializer;